Move tests needing server to IntegrationTests.
Closes #1472.
