+++ src/test/java/org/springframework/data/couchbase/repository/query/StringN1qlQueryCreatorMockedTests.java
@@ -28,7 +28,6 @@ import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.couchbase.config.AbstractCouchbaseConfiguration;
-import org.springframework.data.couchbase.core.CouchbaseTemplate;
 import org.springframework.data.couchbase.core.convert.CouchbaseConverter;
 import org.springframework.data.couchbase.core.convert.MappingCouchbaseConverter;
 import org.springframework.data.couchbase.core.mapping.CouchbaseMappingContext;
@@ -55,57 +54,16 @@ import org.springframework.expression.spel.standard.SpelExpressionParser;
  * @author Michael Nitschinger
  * @author Michael Reiche
  */
-class StringN1qlQueryCreatorMockedTests extends ClusterAwareIntegrationTests {
+class StringN1qlQueryCreatorMockedTests {
 
 	MappingContext<? extends CouchbasePersistentEntity<?>, CouchbasePersistentProperty> context;
 	CouchbaseConverter converter;
-	CouchbaseTemplate couchbaseTemplate;
 	static NamedQueries namedQueries = new PropertiesBasedNamedQueries(new Properties());
 
 	@BeforeEach
 	public void beforeEach() {
 		context = new CouchbaseMappingContext();
 		converter = new MappingCouchbaseConverter(context);
-		ApplicationContext ac = new AnnotationConfigApplicationContext(Config.class);
-		couchbaseTemplate = (CouchbaseTemplate) ac.getBean(COUCHBASE_TEMPLATE);
-	}
-
-	@Test
-	void createsQueryCorrectly() throws Exception {
-		String input = "getByFirstnameAndLastname";
-		Method method = UserRepository.class.getMethod(input, String.class, String.class);
-
-		CouchbaseQueryMethod queryMethod = new CouchbaseQueryMethod(method,
-				new DefaultRepositoryMetadata(UserRepository.class), new SpelAwareProxyProjectionFactory(),
-				converter.getMappingContext());
-
-		StringN1qlQueryCreator creator = new StringN1qlQueryCreator(getAccessor(getParameters(method), "Oliver", "Twist"),
-				queryMethod, converter, "travel-sample", new SpelExpressionParser(),
-				QueryMethodEvaluationContextProvider.DEFAULT, namedQueries);
-
-		Query query = creator.createQuery();
-		assertEquals(
-				"SELECT META(`travel-sample`).id AS __id, META(`travel-sample`).cas AS __cas, `travel-sample`.* FROM `travel-sample` where `_class` = \"org.springframework.data.couchbase.domain.User\" and firstname = $1 and lastname = $2",
-				query.toN1qlSelectString(couchbaseTemplate.reactive(), User.class, false));
-	}
-
-	@Test
-	void createsQueryCorrectly2() throws Exception {
-		String input = "getByFirstnameOrLastname";
-		Method method = UserRepository.class.getMethod(input, String.class, String.class);
-
-		CouchbaseQueryMethod queryMethod = new CouchbaseQueryMethod(method,
-				new DefaultRepositoryMetadata(UserRepository.class), new SpelAwareProxyProjectionFactory(),
-				converter.getMappingContext());
-
-		StringN1qlQueryCreator creator = new StringN1qlQueryCreator(getAccessor(getParameters(method), "Oliver", "Twist"),
-				queryMethod, converter, "travel-sample", new SpelExpressionParser(),
-				QueryMethodEvaluationContextProvider.DEFAULT, namedQueries);
-
-		Query query = creator.createQuery();
-		assertEquals(
-				"SELECT META(`travel-sample`).id AS __id, META(`travel-sample`).cas AS __cas, `travel-sample`.* FROM `travel-sample` where `_class` = \"org.springframework.data.couchbase.domain.User\" and (firstname = $first or lastname = $last)",
-				query.toN1qlSelectString(couchbaseTemplate.reactive(), User.class, false));
 	}
 
 	@Test
@@ -153,29 +111,4 @@ class StringN1qlQueryCreatorMockedTests extends ClusterAwareIntegrationTests {
 		return new DefaultParameters(method);
 	}
 
-	@Configuration
-	@EnableCouchbaseRepositories("org.springframework.data.couchbase")
-	static class Config extends AbstractCouchbaseConfiguration {
-
-		@Override
-		public String getConnectionString() {
-			return connectionString();
-		}
-
-		@Override
-		public String getUserName() {
-			return config().adminUsername();
-		}
-
-		@Override
-		public String getPassword() {
-			return config().adminPassword();
-		}
-
-		@Override
-		public String getBucketName() {
-			return bucketName();
-		}
-
-	}
 }
